Change math::set::contains implementation to if constexpr
Before the propper overload was chosen by traling "requires" clause on multiple function
overloads. This leads to repeating "requires" parameters
(e.g. is_empty_set). To avoid this, a "if constexpr" is now used. This
shows easily, which cases are covered and what the fall through solution
is. It also avoids any duplicate conditions.